+++ Firmware/Board/v3/Inc/gpio.h
@@ -71,11 +71,11 @@ void MX_GPIO_Init(void);
 /* USER CODE BEGIN Prototypes */
 
 void SetGPIO12toUART();
-void SetupENCIndexGPIO();
 bool GPIO_subscribe(GPIO_TypeDef* GPIO_port, uint16_t GPIO_pin,
     uint32_t pull_up_down,
     void (*callback)(void*), void* ctx);
 void GPIO_unsubscribe(GPIO_TypeDef* GPIO_port, uint16_t GPIO_pin);
+void GPIO_set_to_analog(GPIO_TypeDef* GPIO_port, uint16_t GPIO_pin);
 
 
 /* USER CODE END Prototypes */