+++ Firmware/MotorControl/low_level.h
@@ -39,6 +39,9 @@ void start_adc_pwm();
 void start_pwm(TIM_HandleTypeDef* htim);
 void sync_timers(TIM_HandleTypeDef* htim_a, TIM_HandleTypeDef* htim_b,
         uint16_t TIM_CLOCKSOURCE_ITRx, uint16_t count_offset);
+void start_general_purpose_adc();
+
+float get_adc_voltage(GPIO_TypeDef* GPIO_port, uint16_t GPIO_pin);
 
 void update_brake_current();