+++ Firmware/MotorControl/odrive_main.h
@@ -34,6 +34,8 @@ extern bool user_config_loaded;
 extern uint64_t serial_number;
 extern char serial_number_str[13];
 
+#define ADC_CHANNEL_COUNT 16
+extern uint16_t adc_measurements_[ADC_CHANNEL_COUNT];
 
 #ifdef __cplusplus
 }