+++ Firmware/CHANGELOG.md
@@ -35,6 +35,7 @@ Please add a note of your changes below this heading if you make a Pull Request.
 * **USB Bootloader**
 * `make erase_config` to erase the configuration with an STLink (the configuration can also be erased from within explore_odrive.py, using `odrv0.erase_configuration()`)
 * Travis-CI builds firmware for all board versions and deploys the binaries when a tag is pushed to master
+* General purpose ADC API. See function get_adc_voltage() in low_level.cpp for more detais.
 
 ### Changed
 * Most of the code from `lowlevel.c` moved to `axis.cpp`, `encoder.cpp`, `controller.cpp`, `sensorless_estimator.cpp`, `motor.cpp` and the corresponding header files